+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 
 -   Polygon server creation with initial vertices list breaks session
 -   Player floor location not being remembered
+-   Ruler not showing decimal points
 
 ## [0.20.1] - 2020-05-11